Move `MonotonicClock` into clock.rs
