+++ slatedb/src/clock.rs
@@ -1,12 +1,17 @@
 use std::{
+    cmp,
     sync::{
         atomic::{AtomicI64, Ordering},
         Arc,
     },
-    time::{SystemTime, UNIX_EPOCH},
+    time::{Duration, SystemTime, UNIX_EPOCH},
 };
 
-use crate::utils::{system_time_from_millis, system_time_to_millis};
+use crate::{
+    utils::{system_time_from_millis, system_time_to_millis},
+    SlateDBError,
+};
+use tracing::info;
 
 /// Defines the physical clock that SlateDB will use to measure time for things
 /// like garbage collection schedule ticks, compaction schedule ticks, and so on.
@@ -19,13 +24,13 @@ pub trait SystemClock: Send + Sync {
 ///
 /// In test cases, it is possible to advance the clock manually with
 /// `#[tokio::test(start_paused = true)]` and `tokio::time::sleep`.
-pub(crate) struct DefaultSystemClock {
+pub struct DefaultSystemClock {
     initial_ts: i64,
     initial_instant: tokio::time::Instant,
 }
 
 impl DefaultSystemClock {
-    pub(crate) fn new() -> Self {
+    pub fn new() -> Self {
         let ts_millis = match SystemTime::now().duration_since(UNIX_EPOCH) {
             Ok(duration) => duration.as_millis() as i64, // Time is after the epoch
             Err(e) => -(e.duration().as_millis() as i64), // Time is before the epoch, return negative
@@ -89,5 +94,67 @@ impl LogicalClock for DefaultLogicalClock {
         let current_ts = system_time_to_millis(self.inner.now());
         self.last_ts.fetch_max(current_ts, Ordering::SeqCst);
         self.last_ts.load(Ordering::SeqCst)
+    }
+}
+
+/// SlateDB uses MonotonicClock internally so that it can enforce that clock ticks
+/// from the underlying implementation are monotonically increasing
+pub(crate) struct MonotonicClock {
+    pub(crate) last_tick: AtomicI64,
+    pub(crate) last_durable_tick: AtomicI64,
+    delegate: Arc<dyn LogicalClock>,
+}
+
+impl MonotonicClock {
+    pub(crate) fn new(delegate: Arc<dyn LogicalClock>, init_tick: i64) -> Self {
+        Self {
+            delegate,
+            last_tick: AtomicI64::new(init_tick),
+            last_durable_tick: AtomicI64::new(init_tick),
+        }
+    }
+
+    pub(crate) fn set_last_tick(&self, tick: i64) -> Result<i64, SlateDBError> {
+        self.enforce_monotonic(tick)
+    }
+
+    pub(crate) fn fetch_max_last_durable_tick(&self, tick: i64) -> i64 {
+        self.last_durable_tick.fetch_max(tick, Ordering::SeqCst)
+    }
+
+    pub(crate) fn get_last_durable_tick(&self) -> i64 {
+        self.last_durable_tick.load(Ordering::SeqCst)
+    }
+
+    pub(crate) async fn now(&self) -> Result<i64, SlateDBError> {
+        let tick = self.delegate.now();
+        match self.enforce_monotonic(tick) {
+            Err(SlateDBError::InvalidClockTick {
+                last_tick,
+                next_tick: _,
+            }) => {
+                let sync_millis = cmp::min(10_000, 2 * (last_tick - tick).unsigned_abs());
+                info!(
+                    "Clock tick {} is lagging behind the last known tick {}. \
+                    Sleeping {}ms to potentially resolve skew before returning InvalidClockTick.",
+                    tick, last_tick, sync_millis
+                );
+                tokio::time::sleep(Duration::from_millis(sync_millis)).await;
+                self.enforce_monotonic(self.delegate.now())
+            }
+            result => result,
+        }
+    }
+
+    fn enforce_monotonic(&self, tick: i64) -> Result<i64, SlateDBError> {
+        let updated_last_tick = self.last_tick.fetch_max(tick, Ordering::SeqCst);
+        if tick < updated_last_tick {
+            return Err(SlateDBError::InvalidClockTick {
+                last_tick: updated_last_tick,
+                next_tick: tick,
+            });
+        }
+
+        Ok(tick)
     }
 }
